test(crypto): split up crypto tests
