+++ __tests__/unit/crypto/identities/public-key.test.ts
@@ -1,6 +1,7 @@
 import "jest-extended";
 
 import { PublicKey } from "../../../../packages/crypto/src/identities/public-key";
+import { configManager } from "../../../../packages/crypto/src/managers";
 import { data, passphrase } from "./fixture.json";
 
 describe("Identities - Public Key", () => {
@@ -23,6 +24,18 @@ describe("Identities - Public Key", () => {
 
         it("should fail with an invalid public key", () => {
             expect(PublicKey.validate("invalid")).toBeFalse();
+        });
+
+        it("should validate MAINNET public keys", () => {
+            configManager.setConfig(configManager.getPreset("mainnet"));
+
+            expect(PublicKey.validate("02b54f00d9de5a3ace28913fe78a15afcfe242926e94d9b517d06d2705b261f992")).toBeTrue();
+        });
+
+        it("should validate DEVNET public keys", () => {
+            configManager.setConfig(configManager.getPreset("devnet"));
+
+            expect(PublicKey.validate("03b906102928cf97c6ddeb59cefb0e1e02105a22ab1acc3b4906214a16d494db0a")).toBeTrue();
         });
     });
 });
